[batman-adv] Remove compatibility code for	pre 2.6 kernel versions
The earliest version of linux which is compatible to
batman-adv-kernelland is 2.6.20. Their are patches available in debian
for batman-adv-kernelland 0.1 to make it build against linux 2.6.15, but
they cannot be ported to current trunk due to changes to the working
queues. To have "support" for much older versions isn't needed and can
be removed without loosing any functionality.
As result the only remaining LINUX_VERSION_CODE dependend precomiler
code beside compat.h is for netdevice_ops support in soft-interface.c
and result code checking of unregister_chrdev in device.c. For a
possible merge into linux compat.h and all appearances of
LINUX_VERSION_CODE should be removed according to checkpatch.pl.

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>