Staging: batman-adv: fix unlock, make code sparse clean
If the skb failed to copy in the (batman) icmp receive handlers, the function
returned without unlocking the orig_hash_lock first. To prevent this,
the fields needed from the orig_node are now copied, and the copies are
used for sending (just like in other handlers).

Furthermore, find_batman_if() was declared static. The code now appears
to be clean in 'sparse' [1].

Thanks to Andrew Lunn for pointing that out.

[1] http://sparse.wiki.kernel.org/index.php/Main_Page

Signed-off-by: Simon Wunderlich <siwu@hrz.tu-chemnitz.de>