batman-adv: fix endless loop on older kernels
Older kernels (>2.6.23) do not provide the cancel_delayed_work_sync()
function and were redefined to cancel_rearming_delayed_work() which
caused an endless loop, since it waited for the workqueue to re-arm
itself. The batman-adv packet queue would not re-arm itself, hence
it blocked the whole module / kernel in a busy loop.

Signed-off-by: Marek Lindner <lindner_marek@yahoo.de>
Acked-by: Sven Eckelmann <sven.eckelmann@gmx.de>