batman-adv: Only try to add device folders to debugfs if compiled in
Linux allows to call debugfs related functions even when it is not
compiled in. The return codes will always be failure return codes and we
would fail to load the batman-adv module even when debugfs isn't a hard
dependency.

We can simply return a success error code when the debugfs stub returned
a error code and the kernel was compiled without debugfs support.

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>