batman-adv: bonding and interface alternating
This patch adds interface alternating to the new bonding feature. By
default, we now try to avoid forwarding packets on the receiving 
interface, instead choosing alternative interfaces. This feature 
works only on nodes which have multiple interfaces connected to the 
mesh. This approach should reduce problems of the half-duplex nature 
of WiFi Hardware and thus increase performance. 

The patch also includes a bug fix for the interference detection.
Neighbors which are not (yet) candidates should not be considered for
the interference.

Signed-off-by: Simon Wunderlich <siwu@hrz.tu-chemnitz.de>
Acked-by: Marek Lindner <lindner_marek@yahoo.de>
---
Note that unlike the first bonding/alternating patch, the interface
alternating is now activated by default. We have concluded at the
WBMv3 that this is generally a good idea.

We also did tests on dual radio meshnodes, and the link appeared
much more stable (steady 25 Mbit/s, instead of switching between
15 and 25 Mbit/s).