batman-adv: Add last-seen in originator table
Right now, there is no easy/intuitive way to find out whether a node
might have vanished. This commit adds the time when a node was last seen
to the originator table output, so that a common user is able to tell
whether a node might have gone without having to wait PURGE_TIMEOUT
seconds until that node gets "garbe-collected".
It also puts the the versioning information in an extra line, as
the first one of this debug output would otherwise get too long.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>