batman-adv: Don't call unregister_netdev with locked rtnl semaphore
We currently try to call unregister_netdev when we can destroy a softif
after all corresponding hard-interfaces announced that they will be
removed.

This will be done when we receive a hard_if_event which already takes
the rtnl semaphore and thus we try to get it again in unregister_netdev.
This results in a deadlock. This call to unregister_netdev cannot easily
replaced by unregister_netdevice, because other parts of the batman-adv
module still call that code indirectly without holding the rtnl
semaphore.

                    (needs rtln_unlocked)
                     unregister_netdev
                             ^
                             |
                       softif_destroy
                             ^
                             |
                  hardif_disable_interface
                 ^           ^
                /            |
store_mesh_iface   hardif_remove_interface
(rtln_unlocked)              ^            ^
                             |             \
                 hardif_remove_interfaces   hard_if_event
                             ^              (rtln_locked)
                             |
                        batman_exit
                      (rtln_unlocked)

A consistent workaround is to change store_mesh_iface and
hardif_remove_interfaces to call rtnl_lock before they call the
mentioned child function, release the semaphore afterwards and change
unregister_netdev in softif_destroy to unregister_netdevice.

Reported-by: Kazuki Shimada <zukky@bb.banban.jp>
Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>