batman-adv: Include compat.h only once
compat.h is included in all files which may need newer functionality
than the target kernel has. This include is may forgotton when new
changes are made to the sources.

The compat.h can also be included in main.h which is included by all
source files. In theory also all header files must include main.h. This
is currently not and hopefully will never be necessary.

It is also important that all header files which declare functions which
are redefined in compat.h to be included before compat.h is included.

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>