batman-adv: Don't include bat_sysfs.h in compat.h
bat_sysfs.h is included in compat.h to provide bat_attribute which is
used in bat_wrapper_show and bat_wrapper_store. This could create a
include loop which prevents that functions or datastructures defined
elsewhere cannot be found in bat_sysfs.h when batman-adv gets compiled
as external module, but are available when it is used as in-kernel
module.

Signed-off-by: Sven Eckelmann <sven@narfation.org>