batman-adv: Force packed placment of tt_change
Every data structure which is used to send data over the wire must be
fixed in layout to prevent incompatibility between different
architectures. Those incompatibilities could be generated by extra bytes
between members and at the end of the structure.

tt_changes_fill_buffer and tt_len use that structure to copy the correct
amount of data from an internal structure to the actual packet. We must
ensure that this packet buffer only has data which can be interpreted by
any architecture.

Signed-off-by: Sven Eckelmann <sven@narfation.org>