batman-adv: do_bcast has to be true for broadcast packets only
corrects a critical bug of the GW feature. This bug made all the unicast
packets destined to a GW to be sent as broadcast. This bug is present even if
the sender GW feature is configured as OFF. It's an urgent bug fix and should
be committed as soon as possible.

This was a regression introduced by 17852faeb269bc72f6ff542e21c0b3056fb99c48

Signed-off-by: Antonio Quartulli <ordex@autistici.org>