batman-adv: Remove non-kernel parts from Makefile.kbuild
The standalone batman-adv module is sent to the linux kernel maintainers to be
merged by them. A batman-adv developer has to manually merge the changes since
the last submission and apply it on top of the current linux development tree.
The standalone sources have to provide an additional layer that increases the
compatibility to kernels and to give a better user experience. This layer will
not be sent to the kernel maintainers and has to manually removed each time.

To reduce this work, the Makefile.kbuild should be cleaned up to be only the
vanilla version of the Makefile in the kernel. The Makefile in the standalone
sources is used to provide most of the additional layer.

Signed-off-by: Sven Eckelmann <sven@narfation.org>