batman-adv: Allow to enable features without editing Makefile
The user should not be encouraged to edit sources to prevent inconsistent
changes. Upcoming features will not only use preprocessor variables to enable
or disable itself, but also need to compile specific files at the same time.
Therefore, it is safer to provide make parameters which can enable everything
necessary for a specific option and look similar to the .config options of the
kernel.

It is not possible to disable features enabled by the target kernel because the
include/generated/autoconf.h of the kernel sources isn't regenerated by a
module build. Hence, it is better not to disable features in the makefile by
default because the preprocessor will not have the same information. This can
be changed later by a more complex configuration system that generates an own
compat_autoconf.h that is included by compat.h. This header would have to
define/undef all preprocessor variables used to configure batman-adv features.

Signed-off-by: Sven Eckelmann <sven@narfation.org>