batman-adv: Split README in kernel and external part
The README is shipped in a stripped down version as part of the kernel under
Documentation/networking/batman-adv.txt. The readme has to be manually merged
each time it gets prepared for submission to the networking subsystem
maintainer after the README was modified.

This extra work is reduced by splitting it in a README which contains the same
data as available in the kernel and README.external which contains the
information that only affect the external module.

Signed-off-by: Sven Eckelmann <sven@narfation.org>