batman-adv: Prefix bat_sysfs local static functions with batadv_
All non-static symbols of batman-adv were prefixed with batadv_ to avoid
collisions with other symbols of the kernel. Other symbols of batman-adv
should use the same prefix to keep the naming scheme consistent.

This was actually part of patch used to create the commit
98881e49493eb1a6ae2330d861386432064b2222.

Signed-off-by: Sven Eckelmann <sven@narfation.org>

This is the alternative version of this patch... that keeps the functionality
for ELP

 bat_sysfs.c |    9 +++++----
 1 file changed, 5 insertions(+), 4 deletions(-)