batman-adv: send ROAMING_ADV once
If a client roaming has already been advertised, the node should
prevent it from doing the same more than once. To achieve this, the
node has to check the ROAM flag on the global client: if this is set
already, then the ROAMING_ADV for this client has already been sent.

This should be merged with changes done by ("batman-adv: roaming
handling mechanism redesign")

Signed-off-by: Antonio Quartulli <ordex@autistici.org>
Signed-off-by: Marek Lindner <lindner_marek@yahoo.de>