batman-adv: fix compat macro for kstrtou32
The macro previously introduced in the compat code to replace
kstrtou32() may not work properly in case of a too large
user input.

Fix it by using a temporary variable of the proper length
and then cast the result.

Possibly return -ERANGE is the result is too big to fix a 32bit
long variable.

Cc: Matthias Schiffer <mschiffer@universe-factory.net>
Signed-off-by: Antonio Quartulli <antonio@meshcoding.com>
Signed-off-by: Marek Lindner <mareklindner@neomailbox.ch>