batman-adv: Automatically create nested kfree_rcu helper functions
The kfree_rcu compat helper for old kernels require a special function that
knows how to calculate the offsets to the actual data pointer to call kfree.
These can either be manually provided or the GCC extension for nested functions
can be used to automatically create a local function using a macro.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Signed-off-by: Marek Lindner <mareklindner@neomailbox.ch>