batman-adv: Use kernel-like folder structure
The person submitting the patches to the Networking maintainer has to take
each patch and move files around and add some files to make it work in the
actual kernel tree. Having all files included and already  in the correct
kernel tree layout should make this a lot easier.

The only files not part of the kernel tree are compat-include and every
regular file in the top level directory.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Signed-off-by: Marek Lindner <mareklindner@neomailbox.ch>