batman-adv: Don't redefine EPOLL*
The commit 7e040726850a ("eventpoll.h: add missing epoll event masks") was
backported to a couple of older kernels - including 4.4. The build against
such kernel sources creates a lot of extra warnings about redefinitions of
the different EPOLL* variables.

Just disable the definition of these variables when they are already
defined and drop the ones which are not used by batman-adv.

Reported-by: Andre Kasper <andre.kasper@gmx.de>
Signed-off-by: Sven Eckelmann <sven@narfation.org>