batman-adv: Fix colliding redefinition of file_dentry
Linux >=4.4.8 have the function file_dentry already defined while Linux 4.5
only got it with v4.5.2. Older stable kernel versions might also have
gotten this function. Maintaining a full list of kernel versions which have
this functionality is too much of a burden and thus the calls to this
function are simply replaced with calls to a new private function
batadv_file_dentry.

Signed-off-by: Sven Eckelmann <sven@narfation.org>