batman-adv: Drop workaround for Debian's make-kpkg
The make-kpkg is a Debian specific tool to create a debian kernel package.
It is not really maintained anymore and also not used to build the official
Debian kernel packages.

The correct way to have a kernel package under Debian is either to use the
offical packages from the Debian main repository or to build them using the
Linux make target deb-pkg.

There is no reason to ship this workaround when Debian isn't even shipping
make-kpkg with Debian buster and also not in the current Debian bullseye
development snapshot.

Signed-off-by: Sven Eckelmann <sven@narfation.org>