batman-adv: compat: Fix prandom.h include for older LTS kernels
The longterm stable kernels 4.4.x and 4.9.x were missing the
linux/prandom.h for a long time but the commit c0842fbc1b18 ("random32:
move the pseudo-random 32-bit definitions to prandom.h") was recently
backported. Due to this change, it was no longer possible for batman-adv to
find the prandom_u32_max definition. This resulted in a build error:

  net/batman-adv/bat_iv_ogm.c:284:18: error: undefined identifier 'prandom_u32_max'
  net/batman-adv/bat_iv_ogm.c:292:43: error: undefined identifier 'prandom_u32_max'
  net/batman-adv/bat_iv_ogm.c: In function ‘batadv_iv_ogm_emit_send_time’:
  net/batman-adv/bat_iv_ogm.c:284:11: error: implicit declaration of function ‘prandom_u32_max’ [-Werror=implicit-function-declaration]
    msecs += prandom_u32_max(2 * BATADV_JITTER);

Signed-off-by: Sven Eckelmann <sven@narfation.org>