batman-adv: split DAT cache into DAT cache and DAT DHT
Similar to the translation table split the (mostly) locally maintained
DAT cache and foreignly maintained DAT DHT concepts into two separate
tables.

This eases the monitoring and debugging regarding the origin of DAT
entries. And allows to apply differing properties to entries in the DAT
cache and DAT DHT in the future.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>
Signed-off-by: Sven Eckelmann <sven@narfation.org>