batman-adv: Annotate ref_put helper which take locks
When an object is released via kref_put, a release function is called which
might take one or more locks. In this case, it is important that the
kref_put function is never called with this locks held. Since the release
function is only called on release of the object, it is better to warn
about this problem in the reference-put helper to see such a problem
earlier.

Signed-off-by: Sven Eckelmann <sven@narfation.org>